refactor:DB関係のディレクトリを再構築
