+++ domain/user.go
@@ -1,4 +1,4 @@
-package model
+package domain
 
 type User struct {
 	UserID   string `json:"user_id"`
